+++ src/snowflake/snowpark/_internal/open_telemetry.py
@@ -20,13 +20,13 @@ target_modules = [
     "udtf.py",
     "udaf.py",
     "functions.py",
+    "stored_procedure.py",
 ]
-registration_modules = ["udf.py", "udtf.py", "udaf.py"]
+registration_modules = ["udf.py", "udtf.py", "udaf.py", "stored_procedure.py"]
 # this parameter make sure no error when open telemetry is not installed
 open_telemetry_found = True
 try:
     from opentelemetry import trace
-    from opentelemetry.trace import Status, StatusCode
 
 except ImportError:
     open_telemetry_found = False
@@ -62,56 +62,50 @@ def open_telemetry_context_manager(func, dataframe):
 
 
 @contextmanager
-def open_telemetry_udf_context_manager(func, parameters):
+def open_telemetry_udf_context_manager(
+    registration_function,
+    func=None,
+    handler=None,
+    func_name=None,
+    handler_name=None,
+    name=None,
+    file_path=None,
+):
     # trace when required package is installed
     if open_telemetry_found:
-        class_name = func.__qualname__
-        name = func.__name__
+        class_name = registration_function.__qualname__
+        span_name = registration_function.__name__
         tracer = trace.get_tracer(extract_tracer_name(class_name))
-        with tracer.start_as_current_span(name) as cur_span:
+        with tracer.start_as_current_span(span_name) as cur_span:
             try:
                 # first try to get func if it is udf, then try to get handler if it is udtf/udaf, if still None, means it is
                 # loading from file
-                udf_func = (
-                    parameters.get("func")
-                    if parameters.get("func")
-                    else parameters.get("handler")
-                )
+                udf_func = func if func else handler
                 # if udf_func is not None, meaning it is a udf function or udf handler class, get handler_name from it, otherwise find
                 # function name or handler name from parameter
                 handler_name = (
                     udf_func.__name__
                     if udf_func
-                    else (
-                        parameters.get("func_name")
-                        if parameters.get("func_name")
-                        else parameters.get("handler_name")
-                    )
+                    else (func_name if func_name else handler_name)
                 )
                 if cur_span.is_recording():
                     # store execution location in span
                     filename, lineno = context_manager_code_location(
-                        inspect.stack(), func
+                        inspect.stack(), registration_function
                     )
                     cur_span.set_attribute("code.filepath", f"{filename}")
                     cur_span.set_attribute("code.lineno", lineno)
-                    cur_span.set_attribute(
-                        "snow.executable.name", parameters.get("name")
-                    )
+                    cur_span.set_attribute("snow.executable.name", name)
                     cur_span.set_attribute("snow.executable.handler", handler_name)
-                    cur_span.set_attribute(
-                        "snow.executable.filepath", parameters.get("file_path")
-                    )
+                    cur_span.set_attribute("snow.executable.filepath", file_path)
             except Exception as e:
                 logger.warning(f"Error when acquiring span attributes. {e}")
             finally:
                 try:
                     yield
                 except Exception as e:
-                    cur_span.set_status(Status(StatusCode.ERROR, str(e)))
+                    cur_span.set_status(trace.Status(trace.StatusCode.ERROR, str(e)))
                     raise e
-    else:
-        yield
 
 
 def decorator_count(func):